+++ packages/components/src/radio/Radio.js
@@ -6,27 +6,38 @@ import RadioButton from '../common/RadioButton';
 
 import { useDirection } from '../common/hooks';
 
-const Radio = ({ label, id, disabled, secondary, ...otherProps }) => {
+const Radio = ({ label, id, disabled, avatar, secondary, ...otherProps }) => {
   const { isRTL } = useDirection();
   return (
     <div
-      className={classNames('radio', {
+      className={classNames('radio np-radio', {
         'radio-lg': secondary,
-        'radio--rtl': isRTL,
+        disabled,
       })}
       disabled={disabled}
     >
       <label htmlFor={id}>
-        {/* eslint-enable jsx-a11y/label-has-for */}
-        <RadioButton id={id} disabled={disabled} {...otherProps} />
-        {label}
-        {secondary && <small>{secondary}</small>}
+        <span className={classNames('np-radio-button', { 'p-r-2': !isRTL, 'p-l-2': isRTL })}>
+          <RadioButton id={id} disabled={disabled} {...otherProps} />
+        </span>
+        <span className="np-radio__text">
+          {label}
+          {secondary && <small>{secondary}</small>}
+        </span>
+        {avatar && (
+          <span
+            className={classNames('np-radio__avatar', { 'm-l-auto': !isRTL, 'm-r-auto': isRTL })}
+          >
+            {avatar}
+          </span>
+        )}
       </label>
     </div>
   );
 };
 
 Radio.propTypes = {
+  avatar: PropTypes.element,
   checked: PropTypes.bool,
   disabled: PropTypes.bool,
   id: PropTypes.string,
@@ -38,6 +49,7 @@ Radio.propTypes = {
 };
 
 Radio.defaultProps = {
+  avatar: undefined,
   checked: false,
   disabled: false,
   id: null,