+++ packages/dynamic-flows/src/flow/layoutService/LayoutService.spec.js
@@ -42,14 +42,14 @@ describe('Given a utility service for handling dynamic layouts', () => {
             {
               text: option1.description,
               action: {
-                label: option1.title,
+                title: option1.title,
                 method: 'GET',
                 url: option1.url,
               },
             },
             {
               action: {
-                label: option2.title,
+                title: option2.title,
                 method: 'GET',
                 url: option2.url,
                 disabled: true,
@@ -77,17 +77,21 @@ describe('Given a utility service for handling dynamic layouts', () => {
       const finalStep = {
         type: 'final',
         key: 'thing-final',
-        title: 'We create the thing!',
-        description: 'You now do stuff with the thing',
-        image: '/images/1234.png',
+        details: {
+          action: exitAction,
+          title: 'We create the thing!',
+          description: 'You now do stuff with the thing',
+          image: {
+            url: '/images/1234.png',
+          },
+        },
         success: true,
-        action: exitAction,
       };
 
       const finalLayout = [
         {
           type: 'heading',
-          text: finalStep.title,
+          text: finalStep.details.title,
           size: 'lg',
           margin: 'lg',
           align: 'center',
@@ -98,14 +102,14 @@ describe('Given a utility service for handling dynamic layouts', () => {
           components: [
             {
               type: 'image',
-              url: finalStep.image,
+              url: finalStep.details.image.url,
               margin: 'lg',
             },
           ],
         },
         {
           type: 'paragraph',
-          text: finalStep.description,
+          text: finalStep.details.description,
           align: 'center',
         },
         {
@@ -115,7 +119,7 @@ describe('Given a utility service for handling dynamic layouts', () => {
             {
               type: 'button',
               context: exitAction.type,
-              action: { ...exitAction, label: exitAction.title, title: undefined, type: undefined },
+              action: { ...exitAction, type: 'success' },
             },
           ],
         },
@@ -190,7 +194,7 @@ describe('Given a utility service for handling dynamic layouts', () => {
         {
           type: 'review',
           text: reviewFields.title,
-          definitions: [
+          fields: [
             {
               label: reviewFields.fields[0].title,
               value: reviewFields.fields[0].value,
@@ -216,9 +220,8 @@ describe('Given a utility service for handling dynamic layouts', () => {
               context: submitAction.type,
               action: {
                 ...submitAction,
-                label: submitAction.title,
-                title: undefined,
-                type: undefined,
+                title: submitAction.title,
+                type: 'primary',
               },
             },
             {
@@ -226,9 +229,8 @@ describe('Given a utility service for handling dynamic layouts', () => {
               context: cancelAction.type,
               action: {
                 ...cancelAction,
-                label: cancelAction.title,
-                title: undefined,
-                type: undefined,
+                title: cancelAction.title,
+                type: 'delete',
               },
             },
           ],
@@ -322,26 +324,12 @@ describe('Given a utility service for handling dynamic layouts', () => {
           },
         },
         {
-          type: 'button',
-          context: 'primary',
-          action: {
-            $id: '#submitMyDetails',
-            label: 'Continue',
-            url: '/v3',
-            method: 'POST',
-            disabled: false,
-          },
+          $ref: '#submitMyDetails',
+          type: 'action',
         },
         {
-          type: 'button',
-          context: 'primary',
-          action: {
-            $id: '#submitMyAddress',
-            label: 'Continue',
-            url: '/v3',
-            method: 'POST',
-            disabled: false,
-          },
+          $ref: '#submitMyAddress',
+          type: 'action',
         },
       ];
 
@@ -382,15 +370,8 @@ describe('Given a utility service for handling dynamic layouts', () => {
               },
             },
             {
-              type: 'button',
-              context: 'primary',
-              action: {
-                $id: '#submitMyAddress',
-                label: 'Continue',
-                url: '/v3',
-                method: 'POST',
-                disabled: false,
-              },
+              $ref: '#submitMyAddress',
+              type: 'action',
             },
           ],
         },
@@ -445,15 +426,8 @@ describe('Given a utility service for handling dynamic layouts', () => {
               },
             },
             {
-              type: 'button',
-              context: 'primary',
-              action: {
-                $id: '#submitMyDetails',
-                label: 'Continue',
-                url: '/v3',
-                method: 'POST',
-                disabled: false,
-              },
+              $ref: '#submitMyDetails',
+              type: 'action',
             },
           ],
           right: [
@@ -468,15 +442,8 @@ describe('Given a utility service for handling dynamic layouts', () => {
               },
             },
             {
-              type: 'button',
-              context: 'primary',
-              action: {
-                $id: '#submitMyAddress',
-                label: 'Continue',
-                url: '/v3',
-                method: 'POST',
-                disabled: false,
-              },
+              $ref: '#submitMyAddress',
+              type: 'action',
             },
           ],
         },