+++ packages/dynamic-flows/src/layout/DynamicLayout.js
@@ -15,6 +15,7 @@ import DynamicParagraph from './paragraph';
 import DynamicReview from './review';
 
 import { componentModel } from './models';
+import { convertStepToLayout } from '../flow/layoutService';
 
 const DynamicLayout = (props) => {
   const { components, model, submitted, errors, onModelChange, onAction, onPersistAsync } = props;
@@ -84,6 +85,8 @@ const DynamicLayout = (props) => {
         return (
           <DynamicDecision key={getKey(component)} component={component} onAction={onAction} />
         );
+      case 'final':
+        return <>{convertStepToLayout(component).map(renderComponent)}</>;
       default:
         return <div key={getKey(component)} />;
     }
@@ -100,12 +103,13 @@ DynamicLayout.propTypes = {
   baseUrl: Types.string.isRequired,
   onAction: Types.func.isRequired,
   onModelChange: Types.func.isRequired,
-  onPersistAsync: Types.func.isRequired,
+  onPersistAsync: Types.func,
 };
 
 DynamicLayout.defaultProps = {
   model: null,
   errors: null,
+  onPersistAsync: () => {},
 };
 
 export default DynamicLayout;